Remove extra package.json
