--- example/metro.config.js
@@ -1,11 +1,0 @@
-const { makeMetroConfig } = require('@rnx-kit/metro-config');
-module.exports = makeMetroConfig({
-  transformer: {
-    getTransformOptions: async () => ({
-      transform: {
-        experimentalImportSupport: false,
-        inlineRequires: false,
-      },
-    }),
-  },
-});
