--- example/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "MenuExample",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "android": "react-native run-android",
-    "build:android": "mkdirp dist/res && react-native bundle --entry-file index.js --platform android --dev true --bundle-output dist/main.android.jsbundle --assets-dest dist/res",
-    "build:ios": "mkdirp dist && react-native bundle --entry-file index.js --platform ios --dev true --bundle-output dist/main.ios.jsbundle --assets-dest dist",
-    "ios": "react-native run-ios",
-    "lint": "eslint .",
-    "start": "react-native start",
-    "test": "jest"
-  },
-  "dependencies": {
-    "@react-native-menu/menu": "../",
-    "react": "18.2.0",
-    "react-native": "^0.73.6"
-  },
-  "devDependencies": {
-    "@react-native/babel-preset": "^0.75.0-main",
-    "@babel/core": "^7.20.0",
-    "@babel/preset-env": "^7.20.0",
-    "@babel/runtime": "^7.20.0",
-    "@react-native/eslint-config": "0.74.79",
-    "@react-native/metro-config": "0.73.5",
-    "@rnx-kit/metro-config": "^1.3.15",
-    "@tsconfig/react-native": "^3.0.0",
-    "@types/metro-config": "^0.76.3",
-    "@types/react": "^18.0.24",
-    "@types/react-test-renderer": "^18.0.0",
-    "babel-jest": "^29.2.1",
-    "eslint": "^8.19.0",
-    "jest": "^29.2.1",
-    "mkdirp": "^1.0.0",
-    "prettier": "^2.4.1",
-    "react-native-test-app": "^3.2.9",
-    "react-test-renderer": "18.2.0",
-    "typescript": "4.8.4"
-  },
-  "engines": {
-    "node": ">=16"
-  }
-}
