--- example/react-native.config.js
@@ -1,19 +1,0 @@
-const project = (() => {
-  try {
-    const { configureProjects } = require('react-native-test-app');
-    return configureProjects({
-      android: {
-        sourceDir: 'android',
-      },
-      ios: {
-        sourceDir: 'ios',
-      },
-    });
-  } catch (_) {
-    return undefined;
-  }
-})();
-
-module.exports = {
-  ...(project ? { project } : undefined),
-};
