+++ package.json
@@ -20,15 +20,16 @@
     "!**/__mocks__"
   ],
   "scripts": {
-    "test": "jest",
-    "typescript": "tsc --noEmit",
-    "lint": "eslint \"**/*.{js,ts,tsx}\"",
+    "bootstrap": "yarn && yarn example && yarn pods",
+    "example": "yarn --cwd example",
     "format": "prettier --write \"**/*.{js,ts,tsx}\"",
+    "lint": "eslint \"**/*.{js,ts,tsx}\"",
+    "pods": "cd example && pod-install --quiet",
     "prepare": "bob build",
     "release": "release-it",
-    "example": "yarn --cwd example",
-    "pods": "cd example && pod-install --quiet",
-    "bootstrap": "yarn && yarn example && yarn pods"
+    "start": "cd example && react-native start",
+    "test": "jest",
+    "typescript": "tsc --noEmit"
   },
   "keywords": [
     "react-native",
@@ -45,7 +46,9 @@
   "devDependencies": {
     "@react-native/babel-preset": "^0.73.0",
     "@react-native/eslint-config": "^0.74.0",
+    "@react-native/metro-config": "0.73.5",
     "@release-it/conventional-changelog": "^4.3.0",
+    "@rnx-kit/metro-config": "^1.3.15",
     "@types/jest": "^29.1.2",
     "@types/react": "^18.2.2",
     "@types/react-native": "0.72.2",
@@ -59,6 +62,7 @@
     "react": "18.2.0",
     "react-native": "^0.73.0",
     "react-native-builder-bob": "^0.23.2",
+    "react-native-test-app": "^3.2.9",
     "release-it": "^14.10.0",
     "typescript": "4.8.4"
   },
